add admin page, filter, search

+++ src/main/java/com/example/webeconomy/dto/request/ProductUpdateDto.java
@@ -3,11 +3,14 @@ package com.example.webeconomy.dto.request;
 import java.util.List;
 
 import javax.validation.constraints.Min;
+import javax.validation.constraints.NotBlank;
+
 import lombok.Data;
 
 @Data
 public class ProductUpdateDto {
-	private long categoryId;
+	
+	private int categoryId;
 	private String name;
 	private String description;
 	@Min(value = 1, message = "Price should be a positive number")
